modularize-photo closes #137
